card baca cerita reusable dan rename file
