+++ ceritadongeng/src/App.js
@@ -6,19 +6,16 @@ import LibBuku from './pages/LibBuku';
 import { Route, Routes } from 'react-router-dom';
 import DaftarBacaan from './pages/DaftarBacaan';
 import Profil from './pages/Profil';
-import KelinciKura from './pages/BacaCerita/KelinciKura';
 import Loginn from './components/FormLogin/Login';
 import PhoneSignUp from './components/FormLogin/SiginPhone';
 import Signup from './components/FormLogin/SignUp';
 import { UserAuthContextProvider } from './context';
-import KancilBuaya from './pages/BacaCerita/KancilBuaya';
-import BebekBurukRupa from './pages/BacaCerita/Bebek';
-import KancilHarimau from './pages/BacaCerita/KancilHarimau';
 import AboutUs from './pages/AboutUs';
 import Team from './pages/Team';
 import Faq from './pages/Faq';
 import FormUlasan from './pages/FormUlasan';
 import Legal from './pages/Legal';
+import BacaCerita from './pages/BacaCerita';
 
 function App() {
   return (
@@ -33,10 +30,7 @@ function App() {
             <Route path="/library-buku" element={<LibBuku />} />
             <Route path="/daftar-bacaan" element={<DaftarBacaan />} />
             <Route path="/profil" element={<Profil />} />
-            <Route path="/baca-cerita/kelinci-kura" element={<KelinciKura />} />
-            <Route path="/baca-cerita/kancil-harimau" element={<KancilHarimau />} />
-            <Route path="/baca-cerita/kancil-buaya" element={<KancilBuaya />} />
-            <Route path="/baca-cerita/bebek-buruk-rupa" element={<BebekBurukRupa />} />
+            <Route path="/baca-cerita/:judul" element={<BacaCerita />} />
             <Route path="/about-us" element={<AboutUs />} />
             <Route path="/team" element={<Team />} />
             <Route path="/faq" element={<Faq />} />
