+++ ceritadongeng/src/Firestore.js
@@ -11,14 +11,6 @@ const firebaseConfigg = {
   messagingSenderId: '485479888361',
   appId: '1:485479888361:web:fdda295a57c1ac2c290778',
   measurementId: 'G-SL8J46LSP2'
-  // apiKey : process.env.REACT_APP_APIKEY,
-  // authDomain : process.env.REACT_APP_AUTHDOMAIN,
-  // databaseURL : process.env.REACT_APP_DATABASEURL,
-  // projectId : process.env.REACT_APP_PROJECTID,
-  // storageBucket :process.env.REACT_APP_STORAGEBUCKET,
-  // messagingSenderId : process.env.REACT_APP_MESSAGINGSENDERID,
-  // appId : process.env.REACT_APP_APPID,
-  // measurementId : proces.senv.REACT_APP_MEASUREMENTID
 };
 
 const app = initializeApp(firebaseConfigg);
