+++ ceritadongeng/src/components/LibBukuCard/index.css
@@ -4,7 +4,7 @@
 }
 
 .lib-buku-card .btn-heart {
-    margin-left: 50px;
+    margin-left: 40px;
     background-color: transparent;
     border: none;
 }
