+++ ceritadongeng/src/data/card.js
@@ -4,27 +4,27 @@ export const dataCard = [
     dbaseId: 'Bebek Buruk Rupa',
     title: 'Bebek Buruk Rupa',
     writer: 'Hans Christian Andersen',
-    link: '/BacaCerita/BebekBurukRupa'
+    link: '/baca-cerita/bebek_buruk_rupa'
   },
   {
     id: 'kelinci-dan-kura-kura',
     dbaseId: 'Kelinci dan Kura-Kura',
     title: 'Kelinci dan Kura Kura',
     writer: 'Aesop',
-    link: '/BacaCerita/KelinciKura'
+    link: '/baca-cerita/kelinci_kura'
   },
   {
     id: 'kancil-dan-harimau',
     dbaseId: 'Kancil dan Harimau',
     title: 'Kancil dan Harimau',
     writer: 'Dhanny Larito',
-    link: '/BacaCerita/KancilHarimau'
+    link: '/baca-cerita/kancil_harimau'
   },
   {
     id: 'kancil-dan-buaya',
     dbaseId: 'Kancil dan Buaya',
     title: 'Kancil dan Buaya',
     writer: 'Kathy Morris, John Morris',
-    link: '/BacaCerita/KancilBuaya'
+    link: '/baca-cerita/kancil_buaya'
   }
 ];
