+++ ceritadongeng/src/pages/DaftarBacaan/index.js
@@ -2,7 +2,7 @@ import './index.css';
 import { Container, CardGroup, Col } from 'react-bootstrap';
 import Footer from '../../components/Footer';
 import SectionTitle from '../../components/SectionTitle';
-import CardBacaaan from '../../components/CardDaftarBacaan/DaftarBacaan';
+import CardBacaaan from '../../components/CardDaftarBacaan';
 import NavbarHome from '../../components/Navbar';
 
 const DaftarBacaan = () => {
