--- ceritadongeng/src/pages/bacaCerita/KancilBuaya.js
@@ -1,61 +1,0 @@
-import './SliderBacaCerita.css';
-import { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { IoArrowUndoCircleSharp, IoArrowRedoCircleSharp } from 'react-icons/io5';
-import { RiCloseCircleFill } from 'react-icons/ri';
-import db from '../../firestore';
-import { collection, onSnapshot, query, orderBy } from 'firebase/firestore';
-
-const KancilBuaya = () => {
-  const [current, setCurrent] = useState(0);
-  const [cerita, setCerita] = useState([]);
-  const length = cerita.length;
-
-  useEffect(() => {
-    const q = query(collection(db, 'kancil_buaya'), orderBy('index'));
-    onSnapshot(q, (snapshot) => {
-      setCerita(snapshot.docs.map((doc) => ({ ...doc.data(), id: doc.id })));
-    });
-  }, [cerita]);
-
-  const nextSlide = () => {
-    setCurrent(current === length - 1 ? 0 : current + 1);
-  };
-
-  const prevSlide = () => {
-    setCurrent(current === 0 ? length - 1 : current - 1);
-  };
-
-  return (
-    <div className="slider-baca-cerita">
-      <Link to="/library-buku">
-        <RiCloseCircleFill color="red" className="close-cerita" />
-      </Link>
-      <IoArrowUndoCircleSharp className="left-arrow-baca-cerita" onClick={prevSlide} />
-      <IoArrowRedoCircleSharp className="right-arrow-baca-cerita" onClick={nextSlide} />
-      {cerita.map((slide, index) => {
-        return (
-          <div
-            className={index === current ? 'slide-baca-cerita active' : 'slide-baca-cerita'}
-            key={index}>
-            {index === current && (
-              <div>
-                <img src={slide.img} alt=" " className="image-baca-cerita" />
-                <div className="text-cerita">
-                  <p> {slide.text}</p>
-                </div>
-                <div className="btn-audio">
-                  <audio controls>
-                    <source src={slide.suara} type="audio/mpeg" />
-                  </audio>
-                </div>
-              </div>
-            )}
-          </div>
-        );
-      })}
-    </div>
-  );
-};
-
-export default KancilBuaya;
